move code to where it is used
